Conditional Pareto fronts obtained from Gaussian processes simulations.

The data has the only goal of providing an example of use of vorob_t() and vorob_dev(). It has been obtained by fitting two Gaussian processes on 20 observations of a bi-objective problem, before generating conditional simulation of both GPs at different locations and extracting non-dominated values of coupled simulations.

Usage

CPFs

Format

A data frame with 2967 observations on the following 3 variables.

f1

first objective values.

f2

second objective values.

set

indices of corresponding conditional Pareto fronts.

Source

Mickaël Binois, David Ginsbourger, Olivier Roustant (2015). “Quantifying uncertainty on Pareto fronts with Gaussian process conditional simulations.” European Journal of Operational Research, 243(2), 386–394. doi:10.1016/j.ejor.2014.07.032 .

Examples

data(CPFs)
vorob_t(CPFs, reference = c(2, 200))
#> $threshold
#> [1] 44.14062
#> 
#> $ve
#>               [,1]      [,2]
#>   [1,] -37.5978423 -10.79646
#>   [2,] -37.4804644 -11.53365
#>   [3,] -37.0991315 -13.28225
#>   [4,] -36.8538299 -13.51379
#>   [5,] -35.9143692 -15.88340
#>   [6,] -33.7273974 -16.62457
#>   [7,] -33.6535102 -16.82444
#>   [8,] -33.1511911 -18.74396
#>   [9,] -32.9692979 -18.74794
#>  [10,] -31.9646706 -18.75199
#>  [11,] -31.7587126 -19.01094
#>  [12,] -31.5938459 -19.25569
#>  [13,] -31.0546051 -19.27765
#>  [14,] -29.9981435 -19.42665
#>  [15,] -29.7032739 -19.97302
#>  [16,] -28.8450841 -19.99701
#>  [17,] -28.7662560 -20.14155
#>  [18,] -28.7604414 -20.21210
#>  [19,] -28.6222054 -20.29300
#>  [20,] -28.6081935 -20.35386
#>  [21,] -27.5789696 -20.56612
#>  [22,] -27.4631376 -20.65248
#>  [23,] -26.6204447 -20.92707
#>  [24,] -25.7823451 -20.98034
#>  [25,] -25.7284367 -21.07042
#>  [26,] -25.2472705 -21.10899
#>  [27,] -25.2011926 -21.21124
#>  [28,] -25.1393615 -21.26280
#>  [29,] -24.6906978 -21.29381
#>  [30,] -24.4908478 -21.41502
#>  [31,] -23.5137561 -21.67388
#>  [32,] -23.3115825 -21.77622
#>  [33,] -23.0923337 -21.80502
#>  [34,] -22.7120508 -21.88500
#>  [35,] -21.2422296 -21.88748
#>  [36,] -20.7717958 -21.98188
#>  [37,] -20.5216678 -22.06040
#>  [38,] -19.4227481 -22.06195
#>  [39,] -19.4154587 -22.32430
#>  [40,] -19.1464967 -22.38983
#>  [41,] -18.7563892 -22.66689
#>  [42,] -18.5288873 -22.74792
#>  [43,] -17.2773218 -22.78154
#>  [44,] -15.3792174 -22.87815
#>  [45,] -15.0369845 -22.87827
#>  [46,] -14.1304836 -22.89749
#>  [47,] -13.6576356 -22.91511
#>  [48,] -13.4737509 -23.10140
#>  [49,] -12.5828793 -23.20047
#>  [50,] -12.5806182 -23.25464
#>  [51,] -12.4567703 -23.37415
#>  [52,] -12.1855012 -23.44729
#>  [53,] -11.8617280 -23.44904
#>  [54,] -11.7554838 -23.47981
#>  [55,] -11.2052699 -23.48659
#>  [56,]  -9.5701487 -23.50222
#>  [57,]  -9.1922152 -23.56067
#>  [58,]  -9.1141942 -23.59717
#>  [59,]  -8.7050441 -23.69267
#>  [60,]  -8.4556581 -23.88606
#>  [61,]  -8.2849670 -24.06901
#>  [62,]  -8.2392877 -24.09147
#>  [63,]  -8.0321385 -24.09929
#>  [64,]  -7.8725159 -24.12909
#>  [65,]  -7.4508325 -24.14180
#>  [66,]  -7.4474946 -24.19948
#>  [67,]  -7.2977411 -24.24551
#>  [68,]  -6.9845013 -24.27163
#>  [69,]  -6.8595590 -24.30992
#>  [70,]  -6.8388461 -24.38878
#>  [71,]  -6.2772159 -24.39190
#>  [72,]  -6.2175064 -24.41195
#>  [73,]  -6.0638418 -24.42932
#>  [74,]  -5.5502105 -24.45534
#>  [75,]  -5.0101700 -24.48061
#>  [76,]  -4.3447716 -24.50950
#>  [77,]  -3.0751063 -24.70106
#>  [78,]  -3.0329554 -24.81540
#>  [79,]  -2.7721849 -24.88356
#>  [80,]  -2.2829770 -24.95552
#>  [81,]  -1.9772500 -24.98507
#>  [82,]  -1.8284903 -25.09108
#>  [83,]  -1.7979003 -25.19219
#>  [84,]  -1.7099778 -25.27572
#>  [85,]  -0.3551225 -25.45475
#>  [86,]   1.4752205 -25.47433
#>  [87,]   1.6394106 -25.65990
#>  [88,]   2.1632632 -25.95665
#>  [89,]   3.3794879 -26.09472
#>  [90,]   3.5055396 -26.15303
#>  [91,]   3.8248183 -26.24308
#>  [92,]   4.0332572 -26.24397
#>  [93,]   4.1387790 -26.34159
#>  [94,]   4.3513727 -26.43825
#>  [95,]   6.3208813 -26.56214
#>  [96,]   7.2998707 -26.63191
#>  [97,]   9.3460150 -26.65485
#>  [98,]   9.5386116 -26.94045
#>  [99,]  10.3781974 -26.96751
#> [100,]  11.6804102 -26.96959
#> [101,]  14.4125986 -27.00102
#> [102,]  16.6933277 -27.01095
#> [103,]  17.8683518 -27.10385
#> [104,]  18.2756069 -27.65269
#> [105,]  19.4882347 -27.71717
#> [106,]  20.1816676 -27.89640
#> [107,]  21.3732947 -28.01213
#> [108,]  21.9884321 -28.08156
#> [109,]  22.0195376 -28.09682
#> [110,]  22.0288958 -28.26190
#> [111,]  23.3059424 -28.55590
#> [112,]  23.5153250 -28.56087
#> [113,]  24.7980488 -28.64589
#> [114,]  26.8934652 -28.66955
#> [115,]  27.0974155 -28.72620
#> [116,]  28.8806091 -28.89576
#> [117,]  29.0315353 -28.90785
#> [118,]  29.9601039 -28.93467
#> [119,]  30.5141538 -28.96768
#> [120,]  34.0007655 -29.07697
#> [121,]  34.2778379 -29.18076
#> [122,]  35.0459845 -29.26500
#> [123,]  35.8108691 -29.49194
#> [124,]  38.6075068 -29.80295
#> [125,]  39.6553751 -29.80366
#> [126,]  39.9937466 -29.81892
#> [127,]  41.0272025 -29.82227
#> [128,]  42.0504690 -29.88635
#> [129,]  45.6096523 -30.01872
#> [130,]  47.7046212 -30.02561
#> [131,]  49.2600525 -30.17545
#> [132,]  49.6427641 -30.40698
#> [133,]  53.0626287 -30.43793
#> [134,]  54.7697400 -30.50487
#> [135,]  54.9914959 -30.52316
#> [136,]  56.0850877 -30.71623
#> [137,]  56.8025563 -30.76961
#> [138,]  59.1633411 -30.92437
#> [139,]  59.5340987 -31.05449
#> [140,]  63.7214774 -31.15033
#> [141,]  63.9688708 -31.23838
#> [142,]  64.5288942 -31.31147
#> [143,]  68.2713249 -31.42743
#> [144,]  68.9108617 -31.45392
#> [145,]  69.8039471 -31.48788
#> [146,]  70.2719871 -31.50890
#> [147,]  72.6160142 -31.54263
#> [148,]  73.6628042 -31.54855
#> [149,]  73.9591043 -31.60323
#> [150,]  74.6389291 -31.64608
#> [151,]  75.2462156 -31.66217
#> [152,]  75.7353427 -31.70506
#> [153,]  76.9137031 -31.77816
#> [154,]  77.4465275 -31.80060
#> [155,]  77.5418527 -31.81632
#> [156,]  78.9251438 -31.84611
#> [157,]  79.5251958 -31.87112
#> [158,]  80.2261408 -31.87627
#> [159,]  81.1123543 -31.88464
#> [160,]  82.0887769 -31.89967
#> [161,]  82.2917624 -32.02414
#> [162,]  83.5546447 -32.04372
#> [163,]  84.1241591 -32.11488
#> [164,]  85.1042890 -32.29158
#> [165,]  85.9680606 -32.42294
#> [166,]  87.5788795 -32.43134
#> [167,]  88.3184250 -32.47432
#> [168,]  89.3680488 -32.47811
#> [169,]  90.9786234 -32.49957
#> [170,]  91.2025371 -32.52547
#> [171,]  92.6006520 -32.53529
#> [172,]  94.4332216 -32.58076
#> [173,]  95.4272615 -32.64220
#> [174,]  96.4968088 -32.65060
#> [175,]  97.8282433 -32.67302
#> [176,]  98.7716071 -32.71148
#> [177,] 100.5973901 -32.74339
#> [178,] 101.0167180 -32.84149
#> [179,] 102.0021813 -32.85672
#> [180,] 102.7220666 -32.87607
#> [181,] 103.0142379 -32.89849
#> [182,] 104.7393644 -32.93105
#> [183,] 104.8689570 -32.95066
#> [184,] 105.5736765 -33.07372
#> [185,] 106.4822534 -33.08142
#> [186,] 106.9804743 -33.14618
#> [187,] 107.2773119 -33.18751
#> [188,] 108.7657769 -33.20285
#> [189,] 109.0124631 -33.21556
#> [190,] 111.2330616 -33.24898
#> [191,] 111.4303834 -33.26963
#> [192,] 111.5961914 -33.28785
#> [193,] 117.1455263 -33.31951
#> [194,] 117.5839258 -33.37224
#> [195,] 117.7582125 -33.37834
#> [196,] 118.2947053 -33.38737
#> [197,] 118.5338082 -33.43432
#> [198,] 119.5898708 -33.44875
#> [199,] 122.0554270 -33.45548
#> [200,] 123.9807878 -33.48766
#> [201,] 124.4892353 -33.50264
#> [202,] 124.8585753 -33.51050
#> [203,] 127.1357276 -33.51250
#> [204,] 128.3280870 -33.57118
#> [205,] 129.3665563 -33.57142
#> [206,] 129.9706893 -33.61145
#> [207,] 131.7867643 -33.61878
#> [208,] 134.6726692 -33.66297
#> [209,] 135.3907127 -33.67311
#> [210,] 138.4449707 -33.69634
#> [211,] 138.5130531 -33.69958
#> [212,] 139.4607302 -33.73302
#> [213,] 168.0747723 -33.76486
#> 
#> $avg_hyp
#> [1] 8943.333
#>