moocore: Core Mathematical Functions for Multi-Objective Optimization

Fast implementations of mathematical operations and performance metrics for multi-objective optimization, including filtering and ranking of dominated vectors according to Pareto optimality, hypervolume metric, C.M. Fonseca, L. Paquete, M. López-Ibáñez (2006) doi:10.1109/CEC.2006.1688440 , epsilon indicator, inverted generational distance, computation of the empirical attainment function, V.G. da Fonseca, C.M. Fonseca, A.O. Hall (2001) doi:10.1007/3-540-44719-9_15 , and Vorob'ev threshold, expectation and deviation, M. Binois, D. Ginsbourger, O. Roustant (2015) doi:10.1016/j.ejor.2014.07.032 , among others.

See also

Author

Maintainer: Manuel López-Ibáñez manuel.lopez-ibanez@manchester.ac.uk (ORCID)

Other contributors:

Carlos Fonseca [contributor]

Luís Paquete [contributor]

Andreia P. Guerreiro [contributor]

Mickaël Binois [contributor]

Michael H. Buselli (AVL-tree library) [copyright holder]

Wessel Dankers (AVL-tree library) [copyright holder]

NumPy Developers (RNG and ziggurat constants) [copyright holder]

Jean-Sebastien Roy (mt19937 library) [copyright holder]

Makoto Matsumoto (mt19937 library) [copyright holder]

Takuji Nishimura (mt19937 library) [copyright holder]